requirements.txt
pyserial
streamlit
numpy==1.19.3